requirements.txt
openai
feedparser
newspaper3k
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
tiktoken
